Fix goroutine and channel leak
The `WatchDeviceChanges` function creates a goroutine that watches for any change notifications coming from the underlying Device. The result of `b.Parse` is sent to the channel `ch` using a blocking send. Because of this, the caller to `WatchDeviceChanges` must continuously empty this channel otherise the goroutine will simply hang on this line, meaning it will never exit and result in a goroutine leak.  So the caller code looks something like this:

```go
	ch, err := b.WatchDeviceChanges(ctx)
	if err != nil {
        ...
    }

	// Make sure to empty the channel
	go func() {
		for range ch {}
	}()
```

This brings us onto the second issue: the channel `ch` is only closed when the context `ctx` is done. However, the goroutine can also exit due to `b.propchanged` sending a `nil`, which occurs when `UnwatchDeviceChanges` is called. In this case, `ch` is never closed, which means the caller to `WatchDeviceChanges` that is emptying `ch` will never quit, again resulting in a goroutine and channel leak. Closing the channel using a `defer` at the beginning of the goroutine resolves this issue.